+++ js/commands/cmd/termsAndPrivacyCmd.js
@@ -1,0 +1,21 @@
+const referenceMsg = require('../msg/referenceMsg')
+
+module.exports = {
+  'termsAndPrivacy': `<b>Terms of Service and Privacy Policy</b>
+
+  Teletenor uses both <a href="https://telegram.org">Telegram</a> and <a href="https://tenor.com">Tenor</a> services to work.
+
+  <b>Tenor</b>
+  <a href="https://tenor.com/legal-terms">Terms of Service</a>
+  <a href="https://tenor.com/legal-privacy">Privacy Policy</a>
+
+  <b>Telegram</b>
+  <a href="https://telegram.org/tos">Terms of Service</a>
+  <a href="https://telegram.org/privacy">Privacy Policy</a>
+
+  <b>Urls</b>
+  https://telegram.org/
+  https://tenor.com/
+
+  ${referenceMsg}`
+}